requirements.txt
fastapi==0.115.6
pytz==2024.2
uvicorn==0.34.0
python-dotenv==1.0.1
python-multipart==0.0.20
SQLAlchemy==2.0.37
pyjwt==2.7.0
python-dateutil==2.8.2
pydantic_settings==2.0.2
pydantic==2.10.5